+++ python/coinbase-agentkit/coinbase_agentkit/action_providers/__init__.py
@@ -2,14 +2,22 @@
 
 from .action_decorator import create_action
 from .action_provider import Action, ActionProvider
-from .allora.allora_action_provider import AlloraActionProvider, allora_action_provider
 from .basename.basename_action_provider import (
     BasenameActionProvider,
     basename_action_provider,
 )
 from .cdp.cdp_api_action_provider import CdpApiActionProvider, cdp_api_action_provider
+from .cdp.cdp_evm_wallet_action_provider import (
+    CdpEvmWalletActionProvider,
+    cdp_evm_wallet_action_provider,
+)
+from .cdp.cdp_smart_wallet_action_provider import (
+    CdpSmartWalletActionProvider,
+    cdp_smart_wallet_action_provider,
+)
 from .compound.compound_action_provider import CompoundActionProvider, compound_action_provider
 from .erc20.erc20_action_provider import ERC20ActionProvider, erc20_action_provider
+from .erc721.erc721_action_provider import Erc721ActionProvider, erc721_action_provider
 from .hyperboliclabs.hyperbolic_action_provider import (
     HyperbolicActionProvider,
     hyperbolic_action_provider,
@@ -27,21 +35,26 @@ from .twitter.twitter_action_provider import TwitterActionProvider, twitter_acti
 from .wallet.wallet_action_provider import WalletActionProvider, wallet_action_provider
 from .weth.weth_action_provider import WethActionProvider, weth_action_provider
 from .wow.wow_action_provider import WowActionProvider, wow_action_provider
+from .x402.x402_action_provider import x402_action_provider, x402ActionProvider
 
 __all__ = [
     "Action",
     "ActionProvider",
     "create_action",
-    "AlloraActionProvider",
-    "allora_action_provider",
     "BasenameActionProvider",
     "basename_action_provider",
     "CdpApiActionProvider",
     "cdp_api_action_provider",
+    "CdpEvmWalletActionProvider",
+    "cdp_evm_wallet_action_provider",
+    "CdpSmartWalletActionProvider",
+    "cdp_smart_wallet_action_provider",
     "CompoundActionProvider",
     "compound_action_provider",
     "ERC20ActionProvider",
     "erc20_action_provider",
+    "Erc721ActionProvider",
+    "erc721_action_provider",
     "HyperbolicActionProvider",
     "hyperbolic_action_provider",
     "MorphoActionProvider",
@@ -64,4 +77,6 @@ __all__ = [
     "weth_action_provider",
     "WowActionProvider",
     "wow_action_provider",
+    "x402ActionProvider",
+    "x402_action_provider",
 ]