+++ python/coinbase-agentkit/tests/action_providers/cdp/test_cdp_api_action_provider.py
@@ -35,11 +35,11 @@ def test_supports_network():
         )
         assert (
             provider.supports_network(Network(protocol_family="evm", network_id="base-mainnet"))
-            is False
+            is True
         )
         assert (
             provider.supports_network(Network(protocol_family="evm", network_id="ethereum-mainnet"))
-            is False
+            is True
         )
 
         assert (
@@ -48,7 +48,5 @@ def test_supports_network():
         )
         assert (
             provider.supports_network(Network(protocol_family="svm", network_id="solana-mainnet"))
-            is False
+            is True
         )
-
-        assert provider.supports_network(Network(protocol_family="other")) is False